+++ question/type/kprime/backup/moodle2/backup_qtype_kprime_plugin.class.php
@@ -26,8 +26,6 @@
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-defined('MOODLE_INTERNAL') || die();
-
 /**
  * Provides the backup for qtype_kprime questions.
  */